requirements.txt
# Feature store and model registry
hopsworks

# Resolve city names from (longitude, latitude) coordinates
geopy==2.4.1

# Read weather data. Unpinned version - if we don't update, we won't get the weather data
openmeteo-requests

# Be more efficient when making REST (Http) requests
requests-cache==1.2.0
retry-requests==2.0.0

# ML framework libraries
xgboost==2.0.3
scikit-learn==1.4.1.post1

# Plot charts
matplotlib==3.8.3

# Read keys from .env file
python-dotenv

gitpython
